# docs/api/openapi-mobile.yaml
openapi: 3.1.0
info:
  title: Minutriporcion Patient Mobile API
  description: Patient-facing JSON API under /rest/mobile/patient/**. Requires Auth0
    Bearer JWT (resource server).
  contact:
    name: Minutriporcion
    url: https://minutriporcion.com
  version: "1.0"
servers:
- url: http://localhost
  description: Generated server url
security:
- bearer-jwt: []
...
paths:
...
  /rest/mobile/patient/visits:
    get:
      tags:
      - Mobile
      summary: List patient visits
      description: Returns a paged list of visit summaries for the authenticated patient.
      operationId: listVisits
      parameters:
      - name: page
        in: query
        description: Zero-based page index
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: status
        in: query
        description: Filter by visit status
        required: false
        schema:
          type: string
          enum:
          - SCHEDULED
          - COMPLETED
          - CANCELLED
      - name: from
        in: query
        description: Inclusive start instant (ISO-8601)
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        description: Inclusive end instant (ISO-8601)
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: Paged visit summaries
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePagedResponseVisitSummaryDto"
        "401":
          description: Missing or invalid JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePagedResponseVisitSummaryDto"
        "403":
          description: Patient account not linked to Auth0 sub
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePagedResponseVisitSummaryDto"
  /rest/mobile/patient/visits/{visitId}:
    get:
      tags:
      - Mobile
      summary: Get visit detail
      description: Returns a single visit owned by the authenticated patient.
      operationId: getVisitDetail
      parameters:
      - name: visitId
        in: path
        description: Visit identifier
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Visit detail
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVisitDetailDto"
        "401":
          description: Missing or invalid JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVisitDetailDto"
        "403":
          description: Patient account not linked to Auth0 sub
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVisitDetailDto"
        "404":
          description: Resource not found or not owned by patient
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVisitDetailDto"
...
components:
  schemas:
...
    ApiResponsePagedResponseVisitSummaryDto:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/PagedResponseVisitSummaryDto"
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    PagedResponseVisitSummaryDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/VisitSummaryDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
    VisitSummaryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventDateTime:
          type: string
          format: date-time
        title:
          type: string
        status:
          type: string
          enum:
          - SCHEDULED
          - COMPLETED
          - CANCELLED
        durationMinutes:
          type: integer
          format: int32
        summaryNotes:
          type: string
    ApiResponseVisitDetailDto:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/VisitDetailDto"
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    VisitDetailDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventDateTime:
          type: string
          format: date-time
        title:
          type: string
        status:
          type: string
          enum:
          - SCHEDULED
          - COMPLETED
          - CANCELLED
        durationMinutes:
          type: integer
          format: int32
        summaryNotes:
          type: string
        description:
          type: string
        peso:
          type: number
          format: double
        estatura:
          type: number
          format: double
        imc:
          type: number
          format: double
        indiceGrasaCorporal:
          type: number
          format: double
        nivelPeso:
          type: string
          enum:
          - BAJO
          - NORMAL
          - ALTO
          - SOBREPESO
        sistolica:
          type: integer
          format: int32
        diastolica:
          type: integer
          format: int32
        pulso:
          type: integer
          format: int32
        indiceGlucemico:
          type: integer
          format: int32
        spo2:
          type: number
          format: double
        temperatura:
          type: number
          format: double
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: Auth0 access token with audience https://api.nutriconsultas.test/mobile
      scheme: bearer
      bearerFormat: JWT